Refactor date handling to use java.time API
Replaced legacy java.util and java.text date/time classes with java.time API throughout the app for improved reliability and readability. Updated all fragments, adapters, and data classes to use LocalDate, ZonedDateTime, and DateTimeFormatter. Refactored UserTrainingFragment to use a button instead of a checkbox for saving plans to the diary, and updated related layout. This modernizes date handling and simplifies code, reducing parsing errors and improving maintainability.

+++ app/src/main/java/com/example/fithub/EditGoalsDialogFragment.kt
@@ -28,7 +28,7 @@ class EditGoalsDialogFragment : DialogFragment() {
         setupSpinners()
 
         arguments?.let {
-            etTargetWeight.setText(it.getDouble("targetWeight").toString())
+            etTargetWeight.setText(it.getDouble("targetWeight").toInt().toString())
 
             val goalType = it.getString("goalType")
             val goalIndex = when(goalType) {
